Add Action<IAvnMenu> to AvaloniaNativeMenuExporter
